https://github.com/conan-io/conan-center-index/pull/21671

+++ recipes/tz/all/conanfile.py
@@ -3,7 +3,7 @@ import os
 from conan import ConanFile
 from conan.tools.build import cross_building
 from conan.tools.gnu import Autotools, AutotoolsToolchain
-from conan.tools.files import get, copy, rmdir
+from conan.tools.files import get, copy, rmdir, download
 from conan.tools.layout import basic_layout
 from conan.tools.microsoft import unix_path
 
@@ -16,7 +16,14 @@ class TzConan(ConanFile):
     homepage = "https://www.iana.org/time-zones"
     description = "The Time Zone Database contains data that represent the history of local time for many representative locations around the globe."
     topics = ("tz", "tzdb", "time", "zone", "date")
+    package_type = "application" # This is not an application, but application has the correct traits to provide a runtime dependency on data
     settings = "os", "build_type", "arch", "compiler"
+    options = {
+        "with_binary_db": [True, False],
+    }
+    default_options = {
+        "with_binary_db": False,
+    }
 
     def configure(self):
         self.settings.rm_safe("compiler.libcxx")
@@ -27,16 +34,20 @@ class TzConan(ConanFile):
 
     def package_id(self):
         del self.info.settings.compiler
+        if not self.info.options.with_binary_db:
+            self.info.clear()
 
     def build_requirements(self):
-        self.tool_requires("mawk/1.3.4-20230404")
-        if self.settings_build.os == "Windows":
-            self.win_bash = True
-            if not self.conf.get("tools.microsoft.bash:path", check_type=str):
-                self.tool_requires("msys2/cci.latest")
+        if self.options.with_binary_db:
+            self.tool_requires("mawk/1.3.4-20230404")
+            if self.settings_build.os == "Windows":
+                self.win_bash = True
+                if not self.conf.get("tools.microsoft.bash:path", check_type=str):
+                    self.tool_requires("msys2/cci.latest")
 
     def source(self):
-        get(self, **self.conan_data["sources"][self.version], strip_root=True)
+        get(self, **self.conan_data["sources"][self.version]["sources"], strip_root=True)
+        download(self, **self.conan_data["sources"][self.version]["windows_zones"], filename="windowsZones.xml")
 
     def generate(self):
         tc = AutotoolsToolchain(self)
@@ -49,23 +60,65 @@ class TzConan(ConanFile):
         tc.generate()
 
     def build(self):
-        autotools = Autotools(self)
-        autotools.make(args=["-C", self.source_folder.replace("\\", "/")])
+        if self.options.with_binary_db:
+            autotools = Autotools(self)
+            autotools.make(args=["-C", self.source_folder.replace("\\", "/")])
 
     def package(self):
         copy(self, "LICENSE", dst=os.path.join(self.package_folder, "licenses"), src=self.source_folder)
-        autotools = Autotools(self)
-        destdir = self.package_folder.replace('\\', '/')
-        autotools.install(args=["-C", self.source_folder.replace("\\", "/"), f"DESTDIR={destdir}"])
-        rmdir(self, os.path.join(self.package_folder, "usr", "share", "man"))
-        # INFO: The library does not have a public API, it's used to build the zic and zdump tools
-        rmdir(self, os.path.join(self.package_folder, "usr", "lib"))
+        if self.options.with_binary_db:
+            autotools = Autotools(self)
+            destdir = self.package_folder.replace('\\', '/')
+            autotools.install(args=["-C", self.source_folder.replace("\\", "/"), f"DESTDIR={destdir}"])
+            rmdir(self, os.path.join(self.package_folder, "usr", "share", "man"))
+            # INFO: The library does not have a public API, it's used to build the zic and zdump tools
+            rmdir(self, os.path.join(self.package_folder, "usr", "lib"))
+        else:
+            tzdata = [
+                # This file listing is drawn from the source distribution of the tz database at
+                # https://data.iana.org/time-zones/releases/tzdata2023c.tar.gz. It includes only data
+                # files, and excludes project documentation such as CONTRIBUTING, NEWS, README,
+                # SECURITY, theory.html.
+                "africa",
+                "antarctica",
+                "asia",
+                "australasia",
+                "backward",
+                "backzone",
+                "calendars",
+                "checklinks.awk",
+                "checktab.awk",
+                "etcetera",
+                "europe",
+                "factory",
+                "iso3166.tab",
+                "leap-seconds.list",
+                "leapseconds",
+                "leapseconds.awk",
+                "northamerica",
+                "southamerica",
+                "version",
+                "ziguard.awk",
+                "zishrink.awk",
+                "zone.tab",
+                "zone1970.tab",
+                # This file is maintained by CLDR and is required to provide a conversion between
+                # windows time zone names and the IANA time zone names. This enables the IANA tzdb
+                # to be used on windows. For more information, see https://cldr.unicode.org/index
+                "windowsZones.xml",
+            ]
+            for data in tzdata:
+                copy(self, data, dst=os.path.join(self.package_folder, "res", "tzdata"), src=self.source_folder)
 
     def package_info(self):
         self.cpp_info.libdirs = []
         self.cpp_info.includedirs = []
         self.cpp_info.frameworkdirs = []
-        self.cpp_info.resdirs = [os.path.join("usr", "share")]
-        self.cpp_info.bindirs = [os.path.join("usr", "bin"), os.path.join("usr", "sbin")]
-        self.buildenv_info.define("TZDATA", os.path.join(self.package_folder, "usr", "share", "zoneinfo"))
-        self.runenv_info.define("TZDATA", os.path.join(self.package_folder, "usr", "share", "zoneinfo"))
+        self.cpp_info.resdirs = ["res"]
+        self.buildenv_info.define("TZDATA", os.path.join(self.package_folder, "res", "tzdata"))
+        self.runenv_info.define("TZDATA", os.path.join(self.package_folder, "res", "tzdata"))
+        if self.options.with_binary_db:
+            self.cpp_info.resdirs = [os.path.join("usr", "share")]
+            self.cpp_info.bindirs = [os.path.join("usr", "bin"), os.path.join("usr", "sbin")]
+            self.buildenv_info.define("TZDATA", os.path.join(self.package_folder, "usr", "share", "zoneinfo"))
+            self.runenv_info.define("TZDATA", os.path.join(self.package_folder, "usr", "share", "zoneinfo"))
